ui: syntax: provide non-cgo version
Arm has no CGo support. When CGO_ENABLED=0, a classic text edit is used.

However, to make public key text edit and "block untunneled traffic"
toggle work, we still need to parse the tunnel configuration. Rather
than reinventing the wheel, the highlighter.c was ported to Go.

Unfortunately, this solution duplicates the highlighter and is not
viable to maintain in the long term.

Should Go ever become available for Windows 10 ARM64, the Go version of
the highlighter can be dropped. Otherwise, if Go version turns out to be
fast enough to be usable, perhaps we discontinue CGo version of the
highlighter and port the remaining parts of syntax-highlighter (RichEdit
support) to Go.

Signed-off-by: Simon Rozman <simon@rozman.si>